cleaned up REST API code

+++ Bookmarkt Flask/bookmarkt.py
@@ -77,11 +77,14 @@ def getAllUsers():
     except:
         print("error occured")
 
-    return jsonify(jsonList)
+    return jsonify(jsonList), 200
 
 
 @app.route('/users/<userID>', methods=["GET"])
 def getSpecificUser(userID):
+
+    if User.query.filter(User.id == userID).first() is None:
+        return "User does not exist", 422
 
     jsonList = []
     try:
@@ -94,7 +97,7 @@ def getSpecificUser(userID):
     except:
         print("error occured")
 
-    return jsonify(jsonList)
+    return jsonify(jsonList[0]), 200
 
 
 @app.route('/users/<userID>/books/all', methods=["GET"])
@@ -125,7 +128,7 @@ def getAllUserBooks(userID):
                 "googleID": book.googleID
             }
 
-    return jsonify(JsonList)
+    return jsonify(JsonList), 200
 
 
 @app.route("/bookinstance/all", methods=["GET"])
@@ -156,10 +159,10 @@ def getAllBookInstances():
                 "googleID": book.googleID
             }
 
-    return jsonify(JsonList)
+    return jsonify(JsonList), 200
 
 
-@app.route("/users/<userID>/books/add", methods=["GET", "POST"])
+@app.route("/users/<userID>/books/add", methods=["POST"])
 def addUserBook(userID):
 
     isbn = request.args.get("isbn", None)
@@ -173,6 +176,7 @@ def addUserBook(userID):
         except Exception as e:
             print(e)
             print("An Error has occurred")
+            return "currentPage value not valid", 422
     else:
         currentPage = 0
 
@@ -184,9 +188,19 @@ def addUserBook(userID):
     if bookshelfID is not None:
         try:
             bookshelfID = int(bookshelfID)
+
+            #checks to see if bookshelf exists and belongs to that user
+            bookshelf = Bookshelf.query.filter(Bookshelf.bookshelfID == bookshelfID).first()
+            if bookshelf is None:
+                return "Bookshelf does not not exist", 422
+
+            if bookshelf.userID != int(userID):
+                return "Bookshelf does not belong to that user", 403
+
         except Exception as e:
             print(e)
             print("An Error has occurred")
+            return "bookshelfID is not valid", 422
 
     newBookInstance = BookInstance(isbn, userID, completed=completed, currentPage=currentPage, bookshelfID=bookshelfID)
     db.session.add(newBookInstance)
@@ -199,10 +213,10 @@ def addUserBook(userID):
         db.session.add(newBook)
         db.session.commit()
 
-    return "added new BookInstance"
+    return "added new BookInstance", 201
 
 
-@app.route('/users/<userID>/books/<bookInstanceID>/edit', methods=["GET", "PUT"])
+@app.route('/users/<userID>/books/<bookInstanceID>/edit', methods=["PUT"])
 def updateBookInstance(userID, bookInstanceID):
 
     userID = int(userID)
@@ -212,10 +226,13 @@ def updateBookInstance(userID, bookInstanceID):
     bookshelfID = request.args.get("bookshelfID", None)
     bookInstance = BookInstance.query.filter(BookInstance.bookInstanceID == bookInstanceID).first()
 
+    if bookInstance is None:
+        return "Book Instance does not exist", 422
+
     # check if bookinstance belongs to that userid
     if bookInstance.userID != userID:
         print("Book instance does not belong to user")
-        return f"Book Instance {bookInstanceID} does not belong to user {userID}"
+        return f"Book Instance {bookInstanceID} does not belong to user {userID}", 403
 
     if currentPage is not None:
         try:
@@ -249,48 +266,54 @@ def updateBookInstance(userID, bookInstanceID):
 
     db.session.commit()
 
-    return f"Edited book instance {bookInstanceID}"
+    return f"Edited book instance {bookInstanceID}", 200
 
 
-@app.route("/bookinstance/delete")
-def deleteUserBook():
+@app.route("/bookinstance/delete", methods=["DELETE"])
+def deleteBookInstance():
 
     bookInstanceID = request.args.get("bookInstanceID", None)
+
+    if BookInstance.query.filter(BookInstance.bookInstanceID == bookInstanceID).first() is None:
+        return "Book Instance does not exist", 422
 
     BookInstance.query.filter(BookInstance.bookInstanceID == bookInstanceID).delete()
     db.session.commit()
 
-    return f"deleted book instance id {bookInstanceID}"
+    return f"deleted book instance id {bookInstanceID}", 200
 
 
-@app.route("/users/<userID>/books/delete", methods=["GET", "POST"])
-def deleteUserBook2(userID):
+@app.route("/users/<userID>/books/delete", methods=["DELETE"])
+def deleteUserBookInstance(userID):
 
     bookInstanceID = request.args.get("bookInstanceID", None)
 
     bookInstance = BookInstance.query.filter(BookInstance.bookInstanceID == bookInstanceID).first()
 
+    if bookInstance is None:
+        return "Book Instance not found", 422
+
     # checks to see if book instance belongs to that user
     if bookInstance.userID != int(userID):
         print("BookInstance does not belong to user")
-        return f"Book Instance {bookInstanceID} does not belong to user {userID}"
+        return f"Book Instance {bookInstanceID} does not belong to user {userID}", 403
 
     BookInstance.query.filter(BookInstance.bookInstanceID == bookInstanceID).delete()
     db.session.commit()
 
-    return f"deleted book instance id {bookInstanceID}"
+    return f"deleted book instance id {bookInstanceID}", 200
 
 
-@app.route("/users/<userID>/books/delete/all", methods=["GET", "POST"])
+@app.route("/users/<userID>/books/delete/all", methods=["DELETE"])
 def deleteAllUserBook(userID):
 
     BookInstance.query.filter(BookInstance.userID == userID).delete()
     db.session.commit()
 
-    return f"deleted all book instanced from user {userID}"
+    return f"deleted all book instances from user {userID}", 200
 
 
-@app.route("/users/<userID>/books/<bookInstanceID>/read", methods=["GET", "POST"])
+@app.route("/users/<userID>/books/<bookInstanceID>/read", methods=["POST"])
 def addReadingSession(userID, bookInstanceID):
 
     currentPage = request.args.get("currentPage", None)
@@ -300,37 +323,44 @@ def addReadingSession(userID, bookInstanceID):
     if bookInstance.userID != int(userID):
         print(bookInstance.userID, bookInstance.book.title, userID)
         print("Book instance does not belong to user")
-        return f"Book Instance {bookInstanceID} does not belong to user {userID}"
+        return f"Book Instance {bookInstanceID} does not belong to user {userID}", 403
 
-    if currentPage is None or timeRead is None:
-        return "not enough arguments"
+    if currentPage is None:
+        return "currentPage missing", 422
+    if timeRead is None:
+        return "timeRead missing", 422
     else:
 
         readingSession = ReadingSession(bookInstanceID, currentPage, timeRead, userID)
         db.session.add(readingSession)
 
+        # changes the current page of the bookinstance object
         bookInstance.currentPage = currentPage
 
         db.session.commit()
 
+        return "add reading session", 201
 
-        return "add reading session"
 
-
-@app.route('/users/add', methods=["GET", "POST"])
+@app.route('/users/add', methods=["POST"])
 def addNewUser():
 
     newUsername = request.args.get("username", None)
     email = request.args.get("email", None)
 
+    if newUsername is None:
+        return "username is missing", 422
+    if email is None:
+        return "email is missing", 422
+
     newUser = User(username=newUsername, email=email)
     db.session.add(newUser)
     db.session.commit()
 
-    return "added new User"
+    return "added new User", 201
 
 
-@app.route("/users/<userID>/delete", methods=["GET", "POST"])
+@app.route("/users/<userID>/delete", methods=["DELETE"])
 def deleteUser(userID):
 
     User.query.filter(User.id == userID).delete()
@@ -339,10 +369,10 @@ def deleteUser(userID):
 
     db.session.commit()
 
-    return f"deleted user {userID}"
+    return f"deleted user {userID}", 200
 
 
-@app.route('/dropTable', methods=["GET"])
+@app.route('/dropTable', methods=["DELETE"])
 def dropDatabase():
 
     db.drop_all()
@@ -351,7 +381,7 @@ def dropDatabase():
 
 
 @app.route("/books/all", methods=["GET"])
-def allBooks():
+def getAllBooks():
 
     jsonList = []
     try:
@@ -366,12 +396,13 @@ def allBooks():
     except Exception as e:
         print(e)
         print("error occurred")
+        return "An error has occured", 400
 
-    return jsonify(jsonList)
+    return jsonify(jsonList), 200
 
 
 @app.route("/users/<userID>/bookshelf/all", methods=["GET"])
-def getAllBookshelves(userID):
+def getAllUserBookshelves(userID):
 
     jsonList = []
     try:
@@ -385,25 +416,54 @@ def getAllBookshelves(userID):
         print(e)
         print("error occurred")
 
-    return jsonify(jsonList)
+    return jsonify(jsonList), 200
 
 
-@app.route("/users/<userID>/bookshelf/add", methods=["GET", "POST"])
+@app.route("/bookshelf/all", methods=["GET"])
+def getAllBookshelves():
+
+    jsonList = []
+    try:
+        for bookshelf in Bookshelf.query.all():
+
+            jsonList.append({
+                "bookshelfID": bookshelf.bookshelfID,
+                "name": bookshelf.name,
+                "userID": bookshelf.userID
+            })
+    except Exception as e:
+        print(e)
+        print("error occurred")
+
+    return jsonify(jsonList), 200
+
+
+@app.route("/users/<userID>/bookshelf/add", methods=["POST"])
 def addNewBookshelf(userID):
 
-    bookshelfName = request.args.get("name", "")
+    bookshelfName = request.args.get("name", None)
+
+    if bookshelfName is None:
+        return "Bookshelf name is empty", 422
 
     newBookshelf = Bookshelf(bookshelfName, userID)
     db.session.add(newBookshelf)
     db.session.commit()
 
-    return "added new bookshelf"
+    return "added new bookshelf", 201
 
 
 @app.route("/users/<userID>/bookshelf/<bookshelfID>", methods=["GET"])
 def getBooksFromBookshelf(userID, bookshelfID):
 
     JsonList = []
+
+    bookshelf = Bookshelf.query.filter(Bookshelf.bookshelfID == bookshelfID).first()
+
+    if bookshelf is None:
+        return "Bookshelf does not exist", 422
+    if bookshelf.userID != int(userID):
+        return "Bookshelf does not belong to that user", 403
 
     for index, instance in enumerate(BookInstance.query.filter(BookInstance.bookshelfID == bookshelfID)):
 
@@ -428,10 +488,10 @@ def getBooksFromBookshelf(userID, bookshelfID):
                 "googleID": book.googleID
             }
 
-    return jsonify(JsonList)
+    return jsonify(JsonList), 200
 
 
-@app.route("/users/<userID>/bookshelf/<bookshelfID>/add", methods=["GET", "PUT"])
+@app.route("/users/<userID>/bookshelf/<bookshelfID>/add", methods=["POST"])
 def addBookToBookshelf(userID, bookshelfID):
 
     bookInstanceID = request.args.get("bookInstanceID")
@@ -439,14 +499,19 @@ def addBookToBookshelf(userID, bookshelfID):
     bookInstance = BookInstance.query.filter(BookInstance.bookInstanceID == bookInstanceID).first()
     bookshelf = Bookshelf.query.filter(Bookshelf.bookshelfID == bookshelfID).first()
 
+    if bookInstance is None:
+        return "Book Instance not found", 404
+    if bookshelf is None:
+        return "Bookshelf not found", 404
+
     # checks to see if bookshelf belongs to that user
     if bookInstance.userID == bookshelf.userID:
         bookInstance.bookshelfID = bookshelfID
         db.session.commit()
-        return f"added book {bookInstanceID} to bookshelf {bookshelfID}"
+        return f"added book {bookInstanceID} to bookshelf {bookshelfID}", 200
     else:
         print("Book Instance userID and Bookshelf UserID do not match")
-        return f"That book does not belong to owner of bookshelf"
+        return f"That book does not belong to owner of bookshelf", 403
 
 
 @app.route("/authors/all", methods=["GET"])
@@ -455,7 +520,7 @@ def getAllAuthors():
     JsonList = []
 
     for index, author in enumerate(Author.query.all()):
-
+        print(author.authorName)
         JsonList.append({
             "authorData": {},
             "books": []
@@ -475,13 +540,16 @@ def getAllAuthors():
                 "googleID": book.googleID
             })
 
-    return jsonify(JsonList)
+    return jsonify(JsonList), 200
 
 
 @app.route("/authors/<authorID>", methods=["GET"])
 def getSpecificAuthor(authorID):
 
     author = Author.query.filter(Author.authorID == authorID).first()
+
+    if author is None:
+        return "Author does not exist", 422
 
     JsonList = [{
         "authorData": {},
@@ -502,7 +570,7 @@ def getSpecificAuthor(authorID):
             "googleID": book.googleID
         })
 
-    return jsonify(JsonList)
+    return jsonify(JsonList), 200
